fix price datatype and add price to order entity and migration script

+++ quarkus-outbox/quarkus-outbox-core/src/main/java/com/vromanyu/dto/CreateOrderRequestDto.java
@@ -1,4 +1,4 @@
 package com.vromanyu.dto;
 
-public record CreateOrderRequestDto(String itemName, String productType, int quantity, long price) {
+public record CreateOrderRequestDto(String itemName, String productType, int quantity, double price) {
 }
