+++ quarkus-outbox/quarkus-outbox-core/src/main/java/com/vromanyu/dto/CreateOrderResponseDto.java
@@ -2,5 +2,5 @@ package com.vromanyu.dto;
 
 import java.time.Instant;
 
-public record CreateOrderResponseDto(String orderUuid, String itemName, String productType, int quantity, long price, Instant createdAt) {
+public record CreateOrderResponseDto(String orderUuid, String itemName, String productType, int quantity, double price, Instant createdAt) {
 }
